backend ready

+++ db/migrate/20180226010901_create_ideas.rb
@@ -2,9 +2,9 @@ class CreateIdeas < ActiveRecord::Migration[5.1]
   def change
     create_table :ideas do |t|
       t.string :content
-      t.float :impact
-      t.float :ease
-      t.float :confidence
+      t.integer :impact
+      t.integer :ease
+      t.integer :confidence
       t.float :average_score
       t.references :user, foreign_key: true
